:muscle: Move `expand()` into `expand.ts`
